:label: [#445] Convert Anchor/Link components to typescript
